+++ app/Models/User.php
@@ -11,4 +11,12 @@ class User extends Authenticatable
 {
     protected $table = 'users';
     protected $guarded = false;
+
+    public function admin() {
+        return $this->hasOne(Admin::class, 'user_id', 'id');
+    }
+
+    public function isAdmin() {
+        return $this->admin()->exists();
+    }
 }