+++ app/Models/UserConfirmation.php
@@ -2,12 +2,9 @@
 
 namespace App\Models;
 
-// use Illuminate\Contracts\Auth\MustVerifyEmail;
-use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Foundation\Auth\User as Authenticatable;
-use Illuminate\Notifications\Notifiable;
+use Illuminate\Database\Eloquent\Model;
 
-class UserConfirmation extends Authenticatable
+class UserConfirmation extends Model
 {
     protected $table = 'users_confirmation';
     protected $guarded = false;